requirements.txt
# Video Processing and Editing
moviepy>=2.0.0
opencv-python>=4.9.0
Pillow>=10.2.0

# Browser Automation for Meeting Capture
playwright>=1.41.0

# Audio Processing
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.1
pyaudio>=0.2.14

# Voice Cloning - Will use API-based solutions (ElevenLabs, Resemble.ai)
# or alternative libraries compatible with Python 3.12
# TTS requires Python <3.12, so we'll use elevenlabs SDK instead
elevenlabs>=1.2.0

# Deep Learning Frameworks (using latest compatible versions)
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Face Detection and Processing
# Note: dlib and face-recognition can be tricky to install
# We'll use opencv and mediapipe as alternatives initially
# face-recognition==1.3.0
# dlib==19.24.2
mediapipe>=0.10.13
insightface>=0.7.3

# LLM Integration
anthropic>=0.40.0
openai>=1.0.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# CLI and Configuration
click>=8.1.0
pyyaml>=6.0.0

# Async Support
aiohttp>=3.9.0

# FAL API (Kling video backend)
fal-client>=0.4.0

# OBS WebSocket
obsws-python>=1.6.0

# Screen Capture
pyautogui>=0.9.54
